% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{answer<-}
\alias{answer<-}
\title{Answer}
\usage{
answer(state) <- value
}
\arguments{
\item{state}{Participant's \code{state} object.}

\item{value}{New value for \code{answer}.}
}
\description{
Updates the \code{answer} slot of the participant's \code{state} object.
The \code{answer} slot contains the participant's last response.
}
