% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_tester.R
\docType{data}
\name{AppTester}
\alias{AppTester}
\title{App tester}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AppTester
}
\description{
This is an R6 class for running automated evaluations of psychTestR tests
using the \code{shinytest} framework.
When you create an \code{AppTester} object,
it spins up a Shiny app and a headless web browser,
which you can then interact with by calling the methods of the
\code{AppTester} object.
}
\details{
\itemize{
\item \code{app <- AppTester$new(dir)}:
creates a new \code{AppTester} object. The argument
\code{dir} should be a directory containing a file \code{app.R},
with \code{app.R} containing your psychTestR test.
\item \code{app$stop()} - stops the daemon and deletes the output directory.
Call this once you're finished with the \code{AppTester} object.
\item \code{app$click(buttons)} - clicks all the buttons with IDs
listed in the character vector \code{buttons}.
\item \code{app$click_next()} - clicks the "Next" button.
\item \code{app$get_ui()} - returns the HTML of the current UI of the psychTestR test.
\item \code{app$get_ui_text()} - extracts displayed text from \code{app$get_ui()}
using some heuristics. If \code{squish = TRUE}, extra whitespace
is removed.
\item \code{app$expect_ui_text(text, squish = TRUE)} - throws an error
if the output of \code{app$get_ui_text(squish = squish)}
is not equal to the argument \code{text}.
\item \code{app$get_locals()} - returns the current local variables.
\item \code{app$get_globals()} - returns the current global variables.
\item \code{app$get_title()} - returns the current title.
\item \code{app$expect_title(text)} - throws an error
if the output of \code{app$get_title()}
is not equal to the argument \code{text}.
\item \code{app$get_problems_info()} - returns the "problems info"
text displayed at the bottom of the page.
\item \code{app$expect_problems_info(text)} - throws an error
if the output of \code{app$get_problems_info()}
is not equal to the argument \code{text}.
\item \code{app$get_results()} - returns the current results for the current participant.
\item \code{app$expect_results(val)} - throws an error
if the output of \code{as.list(app$get_results())}
is not equal to \code{val}.
\item \code{app$clean_output()} - deletes the output directory.
\item \code{app$set_inputs(formA = "Text A", formB = "Text B")} -
passes (for example) the string \code{"Text A"} to the UI
input with ID \code{"formA"} and the string \code{"Text B"}
to the UI input with ID \code{"formB"}.
\item \code{...} - \code{AppTester} objects inherit from \code{\link[shinytest]{ShinyDriver}}
objects, and have access to all of their methods.
}
}
\note{
\code{AppTester} depends on the prior installation of a headless web browser, PhantomJS,
which you can install with \code{\link[shinytest]{installDependencies}}.
}
\keyword{datasets}
