% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{as.timeline}
\alias{as.timeline}
\title{Coerce to timeline}
\usage{
as.timeline(x, ...)
}
\arguments{
\item{x}{Object to coerce.}

\item{...}{Further arguments to pass to \code{\link{new_timeline}()}.}
}
\description{
Coerces objects to the \code{timeline} class.
}
