% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_global_is_null}
\alias{assert_global_is_null}
\title{Assert that a global variable is NULL}
\usage{
assert_global_is_null(key, state)
}
\arguments{
\item{key}{Key of the global variable that should be \code{NULL}.}

\item{state}{The participant's \code{state} object.}
}
\description{
Throws an error if a certain global variable is not \code{NULL}.
}
