% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-demographics.R
\name{basic_demographics_default_intro}
\alias{basic_demographics_default_intro}
\title{Default demographics introduction}
\usage{
basic_demographics_default_intro()
}
\description{
Default introduction to the demographics section.
}
