% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{begin_module}
\alias{begin_module}
\title{Begin module}
\usage{
begin_module(label)
}
\arguments{
\item{label}{Module label (character scalar).}
}
\description{
Returns a code block that begins a psychTestR module.
Modules have their own set of local variables.
They also have identifying labels that are stored alongside
result entries created during the module.
}
