% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{check_dirs}
\alias{check_dirs}
\title{Check directories}
\usage{
check_dirs(opt)
}
\arguments{
\item{opt}{Options list as created by \code{test_options()}.}
}
\description{
Checks that all directories as specified in the test options
are writeable.
}
