% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{checkbox_page}
\alias{checkbox_page}
\title{New checkbox page}
\usage{
checkbox_page(
  label,
  prompt,
  choices,
  subprompt = "",
  labels = NULL,
  trigger_button_text = "Continue",
  failed_validation_message = c(one = "Choose at least one answer!", all =
    "Please check all boxes."),
  force_answer = c("no", "one", "all", "first"),
  javascript = "",
  save_answer = TRUE,
  hide_response_ui = FALSE,
  response_ui_id = "response_ui",
  on_complete = NULL,
  admin_ui = NULL
)
}
\arguments{
\item{label}{(Character scalar) Label for the current page.}

\item{prompt}{Prompt to be displayed above the response choices.
Can be either a character scalar (e.g. "What is 2 + 2?") or an object of
class "shiny.tag", e.g. \code{shiny::tags$p("What is 2 + 2?")}.}

\item{choices}{(Character vector) Choices for the participant.
If unnamed, then these values will be used both for checkbox IDs and for
checkbox labels.
If named, then values will be used for checkbox IDs and names will be used
for checkbox labels.}

\item{subprompt}{(Character scalar) Optional additional text in bold letters
below the prompt.}

\item{labels}{Optional vector of labels for the checkbox choices.
If not \code{NULL}, will overwrite the names of \code{choices}.
This vector of labels can either be a character vector
or a list of Shiny tag objects, e.g. as created by \code{shiny::HTML()}.}

\item{trigger_button_text}{(Character scalar) Text for the trigger button.}

\item{failed_validation_message}{(Named character vector) Text to be displayed
when validation fails, names must match the validation type ("one" or "all"), see "force_answer".}

\item{force_answer}{(Boolean scalar or character) Require at least one checkbox (value "one") or all checkboxes ("all") to be
ticked. If no selection is required set to "no" (default). If boolean, this translate to option "one" (TRUE) or "no" (FALSE).}

\item{javascript}{(Character scalar) JavaScript code to be added for
controlling checkbox behaviour.}

\item{save_answer}{(Boolean scalar) Whether or not to save the answer.}

\item{hide_response_ui}{(Boolean scalar) Whether to begin with the response
interface hidden (it can be subsequently made visible through Javascript,
using the element ID as set in \code{response_ui_id}.
See \code{audio_NAFC_page} for an example.).}

\item{response_ui_id}{(Character scalar) HTML ID for the response user
interface.}

\item{on_complete}{Optional function to execute on leaving the page
(after successful validation).
The argument list should include \code{...},
and any of:
\code{state}, the participant's state object;
\code{answer}, the participant's most recent answer;
\code{input}, the current page's Shiny input object;
\code{session}, the current Shiny session object;
\code{opt}, the test's option list as created by \code{test_options()}.}

\item{admin_ui}{Optional UI component for the admin panel.}
}
\description{
Creates a checkbox page.
}
