% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{clean_session_dir}
\alias{clean_session_dir}
\title{Deletes all session caches that are older than a certain amount.}
\usage{
clean_session_dir(session, opt)
}
\description{
Deletes all session caches that are older than a certain amount.
}
