% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{cleanSessionDir}
\alias{cleanSessionDir}
\title{Deletes all session caches that are older than a certain amount.}
\usage{
cleanSessionDir(session_dir, timeout_min = 60)
}
\description{
Deletes all session caches that are older than a certain amount.
}
