% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{conditional}
\alias{conditional}
\title{Conditional test block}
\usage{
conditional(test, logic)
}
\arguments{
\item{test}{Function to evaluate at run time.
This function must accept the argument \code{...},
and optionally the following named arguments:
\itemize{
\item \code{state}, the participant's state object;
\item \code{input}, the current page's Shiny input object;
\item \code{output}, the current page's Shiny output object;
\item \code{session}, the current Shiny session object;
\item \code{opt}, the test's option list as created by \code{\link{test_options}}.
}}

\item{logic}{Either a single test element, a list of test elements, or a timeline,
which will be displayed conditionally on the basis of the outcome
of \code{test}.}
}
\value{
A list of test elements, or equivalently a timeline, which can be combined
with other test elements or timelines.
}
\description{
This function evalutes a test function at run time to decide
whether to administer a series of test elements to the participant.
If the test returns \code{TRUE}, then the test elements are shown,
otherwise these test elements are skipped.
}
