% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-get-url-params.R
\name{demo.get_url_params}
\alias{demo.get_url_params}
\title{Demo: Getting URL parameters}
\usage{
demo.get_url_params()
}
\description{
Demonstrates how psychTestR can get URL parameters.
}
