% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{demo}
\alias{demo}
\title{Demo}
\usage{
demo(state)
}
\arguments{
\item{state}{The participant's \code{state} object.}
}
\value{
\code{TRUE} if the current session is a demo session,
\code{FALSE} otherwise.
}
\description{
Determines whether the current session is a demo session.
}
