% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-set-url-params.R
\name{demo.set_url_params}
\alias{demo.set_url_params}
\title{Demo: Getting URL parameters}
\usage{
demo.set_url_params()
}
\description{
Demonstrates how psychTestR can set URL parameters.
}
