% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{demo_options}
\alias{demo_options}
\title{Demo options}
\usage{
demo_options(title = "Demo", admin_password = "demo",
  researcher_email = "XXX", demo = TRUE, ...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{test_options}()}.}
}
\description{
Test options list for demo purposes.
}
