% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{display_error}
\alias{display_error}
\title{Display a popup error message}
\usage{
display_error(...)
}
\arguments{
\item{...}{Arguments to be pasted together (without separator) to form
the error message.}
}
\description{
Displays a popup error message and terminates the testing session.
}
