% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{dropdown_page}
\alias{dropdown_page}
\title{Make dropdown list page}
\usage{
dropdown_page(prompt, choices, set_global = NULL,
  alternative_choice = FALSE, alternative_text = "Other (please state)",
  next_button_text = "Next", max_width_pixels = 200, validate = "auto",
  ...)
}
\description{
Creates a page where the response is to be selected from a dropdown list.
}
