% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{elt_save_results_to_disk}
\alias{elt_save_results_to_disk}
\title{Save results to disk (test element version)}
\usage{
elt_save_results_to_disk(complete)
}
\arguments{
\item{complete}{Whether or not the participant
has now completed the test (scalar Boolean).}
}
\description{
Returns a test element that saves the current participant's results to disk.
}
