% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{end_module}
\alias{end_module}
\title{End module}
\usage{
end_module()
}
\description{
Returns a code block that ends a psychTestR module.
Modules have their own set of local variables.
They also have identifying labels that are stored alongside
result entries created during the module.
}
\note{
Usually it is better to call \code{\link{module}} instead.
}
