% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{finish_test_and_give_code}
\alias{finish_test_and_give_code}
\title{Finish test and give code}
\usage{
finish_test_and_give_code(researcher_email)
}
\arguments{
\item{researcher_email}{The researcher's email address
(given in case the participant has any further questions).}
}
\description{
psychTestR logic for finishing the current test
and giving the participant a randomly
generated code that can then be used for organising payment
(e.g. for Amazon Mechanical Turk).
}
