% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{format_new_timeline}
\alias{format_new_timeline}
\title{Format new timeline}
\usage{
format_new_timeline(input, langs)
}
\arguments{
\item{input}{Timeline to format.}

\item{langs}{Character vector of desired languages.}
}
\value{
Formatted timeline.
}
\description{
This function takes a timeline object as input and performs some
postprocessing to check which languages are available
and remove undesired languages.
It is called automatically by \code{\link{new_timeline}};
most users will never need to call this function directly.
}
\keyword{internal}
