% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{format_test_element_list}
\alias{format_test_element_list}
\title{Format test element list}
\usage{
format_test_element_list(input, lang)
}
\arguments{
\item{input}{Either a single test element or a list of either
test elements or timelines.}

\item{lang}{(Character scalar)
The desired language of the output.}
}
\value{
A list of test elements expressed in the desired language.
}
\description{
This function formats its input as a list of test elements
expressed in a given internalisation language.
It is automatically called within \code{\link{new_timeline}};
most users will never need to call this function directly.
}
\keyword{internal}
