% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-demographics.R
\name{get_basic_demographics}
\alias{get_basic_demographics}
\title{Get basic demographics}
\usage{
get_basic_demographics(intro = basic_demographics_default_intro(),
  gender = TRUE, age = TRUE, occupation = TRUE, education = TRUE)
}
\arguments{
\item{intro}{Introductory page to the demographics section.
If \code{NULL}, no introduction is given.}

\item{gender}{Ask the participant's gender? (Boolean)}

\item{age}{Ask the participant's age? (Boolean)}

\item{occupation}{Ask the participant's occupation? (Boolean)}

\item{education}{Ask the participant's education? (Boolean)}
}
\description{
Gets basic demographic information from the participant.
}
