% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{get_global}
\alias{get_global}
\title{Get global variable}
\usage{
get_global(key, state)
}
\arguments{
\item{key}{The variable's key (character scalar).}

\item{state}{The participant's \code{state} object.}
}
\value{
The global variable with key equal to \code{key}.
}
\description{
Gets a global variable from the participant's testing session.
}
