% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{get_local}
\alias{get_local}
\title{Get local variable}
\usage{
get_local(key, state)
}
\arguments{
\item{key}{The variable's key (character scalar).}

\item{state}{The participant's \code{state} object.}
}
\value{
The local variable with key equal to \code{key}.
}
\description{
Gets a local variable from the participant's testing session.
}
