% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{get_results}
\alias{get_results}
\title{Get results}
\usage{
get_results(state, complete, add_session_info = FALSE)
}
\arguments{
\item{state}{The participant's \code{state} object.}

\item{complete}{Whether the participant has now completed the test (Boolean).}

\item{add_session_info}{Whether to add session information (Boolean).}

\item{complete}{Whether the participant completed the test.}
}
\value{
A \code{results} object.
}
\description{
Gets the participants current results, optionally
along with session information.
}
