% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{get_session_info}
\alias{get_session_info}
\title{Get session info}
\usage{
get_session_info(state, complete)
}
\arguments{
\item{state}{The participant's \code{state} object.}

\item{complete}{Whether or not the session is now considered complete
(this only affects the returned session info).}
}
\value{
List containing session information.
}
\description{
Gets information about the current participant's session,
including things like participant ID, time started, and number of restarts.
}
