% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary-results.R
\name{get_supplementary_results_dir}
\alias{get_supplementary_results_dir}
\title{Supplementary results directory}
\usage{
get_supplementary_results_dir(opt)
}
\arguments{
\item{opt}{Options list as created by \code{test_options()}.}
}
\value{
Character scalar identifying the (relative)
path to the supplementary results directory.
}
\description{
Returns the location of the supplementary results directory.
The supplementary results directory provides a place to store
results in addition to the results saved by the psychTestR
function \code{save_results_to_disk}.
The researcher has full control over what to put in the supplementary
results directory.
}
\note{
The supplementary results directory is typically stored
within the main results directory. It will be deleted if the
researcher presses the 'Delete all results' button in the admin panel.
}
