% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{get_url_params}
\alias{get_url_params}
\title{Get URL parameters}
\usage{
get_url_params(state)
}
\arguments{
\item{state}{The participant's \code{state} object.}
}
\description{
Gets the current URL parameters as stored in \code{state}.
}
