% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{i18n}
\alias{i18n}
\title{Translate}
\usage{
i18n(x, html = TRUE, sub = character())
}
\arguments{
\item{x}{Character scalar identifying the term to be translated.}

\item{html}{Boolean; whether or not the translation output should
be parsed as HTML or as plain text (the default is \code{TRUE},
corresponding to HTML parsing).}

\item{sub}{A named character vector or list defining substitutions
to make within the translation output. The text of the translation output
may contain passages such as \code{{{var1}}};
if \code{sub = c(var1 = "Hello")},
then \code{{{var1}}} will be replaced with the text \code{Hello}.}
}
\value{
The translated output, typically a character scalar
or a \code{shiny::HTML()} output.
}
\description{
Translates text into a language determined by the current testing session.
Translation is carried out using a dictionary defined at test creation
(see \code{\link{i18n_dict}}).
The function can typically only be used within a call
to \code{\link{new_timeline}}, which defines the dictionary
in which to look up \code{i18n}.
Calling \code{i18n} outside this context will typically result in an error.
}
