% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\docType{data}
\name{I18N_STATE}
\alias{I18N_STATE}
\title{i18n state}
\format{
An object of class \code{i18n_state} (inherits from \code{R6}) of length 11.
}
\usage{
I18N_STATE
}
\description{
This object captures the psychTestR package's current
internationalisation state.
This internationalisation state determines how calls to \code{\link{i18n}}
are resolved into translations.
Most users should not need to work with this object directly,
and its API should not yet be considered stable.
}
\keyword{internal}
