% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{imports}
\alias{imports}
\alias{as_tibble}
\title{Objects imported from other packages}
\description{
These objects are imported from other packages.
Follow the links below to see their documentation.
}
\details{
\code{\link[tibble]{as_tibble}}
}
