% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{increment_elt_index}
\alias{increment_elt_index}
\title{Low-level setter, see skip_n_pages for skipping pages in general}
\usage{
increment_elt_index(state, by = 1L)
}
\description{
Low-level setter, see skip_n_pages for skipping pages in general
}
