% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.null.or}
\alias{is.null.or}
\title{Is NULL or...?}
\usage{
is.null.or(x, f)
}
\arguments{
\item{x}{Object to check.}

\item{f}{Function to apply.}
}
\description{
Returns \code{TRUE} if \code{x} is either \code{NULL} or \code{f(x)}
is \code{TRUE}.
}
\keyword{internal}
