% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{is.test_element}
\alias{is.test_element}
\title{Is object a test element?}
\usage{
is.test_element(x)
}
\arguments{
\item{x}{Object to check.}
}
\description{
Checks whether an object is a test element.
}
