% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{is.timeline}
\alias{is.timeline}
\title{Is it a timeline object?}
\usage{
is.timeline(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
\code{TRUE} if object is a timeline object, \code{FALSE} otherwise.
}
\description{
Checks whether an object is a timeline object.
}
