% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{is_test_closed}
\alias{is_test_closed}
\title{Is test closed?}
\usage{
is_test_closed(opt)
}
\arguments{
\item{opt}{Test options list as created by \code{test_options()}.}
}
\value{
\code{TRUE} if the test is closed, \code{FALSE} otherwise.
}
\description{
Checks whether the test is closed.
}
