% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{join}
\alias{join}
\title{Join test elements}
\usage{
join(...)
}
\arguments{
\item{...}{Test elements or timelines to combine.}
}
\value{
A list of test elements if the input didn't contain
any timelines, otherwise a timeline.
}
\description{
Joins a series of test elements or timelines into
one list or timeline.
Often one can equivalently use `c(...)` in such situations,
but this approach is safer.
}
