% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{loop_while-deprecated}
\alias{loop_while-deprecated}
\title{Loop}
\arguments{
\item{test}{Test function to execute.
Argument list should include \code{...};
further permitted arguments are:
\itemize{
\item \code{state}, the participant's \code{state} object;
\item \code{input}, the Shiny \code{input} object for the current page;
\item \code{output}, the Shiny \code{output} object for the current page;
\item \code{session}, the Shiny \code{session} object;
\item \code{opt}, test options as created by \code{test_options()}.
}}

\item{logic}{List of psychTestR test elements.}
}
\description{
Repeatedly show a series of test elements to the participant
while a condition is satisfied.
The elements are shown at least once; at the end of each pass,
the test function is checked.
}
\seealso{
\code{\link{psychTestR-deprecated}}
}
\keyword{internal}
