% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-test.R
\name{make_test}
\alias{make_test}
\title{Make psychTestR test}
\usage{
make_test(elts, opt = demo_options(), custom_admin_panel = NULL)
}
\arguments{
\item{elts}{List of test elements defining the test's timeline.}

\item{opt}{Options list as created by \code{test_options()}.}

\item{custom_admin_panel}{This argument supports the construction of a custom
admin panel. The argument should be a function that runs during the
Shiny server function. The argument list of this function should
include \code{...}; it may also include
\code{state}, \code{input}, \code{output}, and \code{session},
all with their traditional meanings from Shiny or psychTestR.
This function should include an expression of the form
\code{output$custom_admin_panel <- shiny::renderUI(...)}
where the code inside \code{renderUI()} defines the UI of the admin panel
using standard Shiny UI vocabulary.
Further documentation for this option should be forthcoming.}
}
\description{
The top-level function that defines a psychTestR test.
This should be the final line of your app script,
typically entitled \code{app.R}.
}
