% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{make_ui_checkbox}
\alias{make_ui_checkbox}
\title{Make checkboxes}
\usage{
make_ui_checkbox(
  label,
  choices,
  subprompt = "",
  labels = NULL,
  trigger_button_text = "Continue",
  javascript = "",
  hide = FALSE,
  id = "response_ui"
)
}
\arguments{
\item{label}{(Character scalar) Label for the current page.}

\item{choices}{(Character vector) Choices for the participant.
If unnamed, then these values will be used both for checkbox IDs and for
checkbox labels.
If named, then values will be used for checkbox IDs and names
will be used for checkbox labels.}

\item{subprompt}{(Character scalar) Optional additional text in bold letters
below the prompt.}

\item{labels}{Optional vector of labels for the NOMC choices.
If not \code{NULL}, will overwrite the names of \code{choices}.
This vector of labels can either be a character vector
or a list of Shiny tag objects, e.g. as created by \code{shiny::HTML()}.}

\item{trigger_button_text}{(Character scalar) Text for the trigger button.}

\item{javascript}{(Character scalar) JavaScript code to be added for
controlling checkbox behaviour.}

\item{hide}{(Boolean scalar) Whether the checkboxes should be hidden
(possibly to be shown later).}

\item{id}{(Character scalar) HTML ID for the div containing the checkboxes.}
}
\description{
Creates HTML code for checkbox response options.
}
