% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{make_ui_NAFC}
\alias{make_ui_NAFC}
\title{Make NAFC buttons}
\usage{
make_ui_NAFC(choices, hidden = FALSE, id = "response_UI",
  arrange_vertically = TRUE)
}
\arguments{
\item{choices}{Character vector of choices for the participant.
If unnamed, then these values will be used both for button IDs
and for button labels.
If named, then values will be used for button IDs and names
will be used for button labels.}

\item{hidden}{Whether the response buttons should be hidden
(possibly to be shown later).}

\item{id}{HTML ID for the div containing the response buttons.}

\item{arrange_vertically}{Whether to arrange the response buttons vertically
(the default) as opposed to horizontally.}
}
\description{
Creates HTML code for n-alternative forced-choice response options.
}
