% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{make_ui_radiobutton_NAFC}
\alias{make_ui_radiobutton_NAFC}
\title{Make NAFC radiobuttons}
\usage{
make_ui_radiobutton_NAFC(
  label,
  choices,
  subprompt = "",
  labels = NULL,
  trigger_button_text = "Continue",
  hide = FALSE,
  id = "response_ui"
)
}
\arguments{
\item{label}{(Character scalar) Label for the current page.}

\item{choices}{(Character vector) Choices for the participant.
If unnamed, then these values will be used both for radiobutton IDs and for
button labels.
If named, then values will be used for button IDs and names
will be used for button labels.}

\item{subprompt}{(Character scalar) Optional additional text in bold letters
below the prompt.}

\item{labels}{Optional vector of labels for the NAFC radiobutton choices.
If not \code{NULL}, will overwrite the names of \code{choices}.
This vector of labels can either be a character vector
or a list of Shiny tag objects, e.g. as created by \code{shiny::HTML()}.}

\item{trigger_button_text}{(Character scalar) Text for the trigger button.}

\item{hide}{(Boolean scalar) Whether the radiobuttons should be hidden (possibly to be shown later).}

\item{id}{(Character scalar) HTML ID for the div containing the radiobuttons.}
}
\description{
Creates HTML code for n-alternative forced-choice response radiobutton options.
}
