% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{NAFC_page}
\alias{NAFC_page}
\title{New NAFC page}
\usage{
NAFC_page(label, prompt, choices, save_answer = TRUE,
  arrange_vertically = length(choices) > 2L, hide_response_ui = FALSE,
  response_ui_id = "response_ui", on_complete = NULL, admin_ui = NULL)
}
\arguments{
\item{prompt}{Prompt to be displayed above the response choices.
Can be either a scalar character (e.g. "What is 2 + 2?")
or an object of class "shiny.tag", e.g. \code{shiny::tags$p("What is 2 + 2?")}.}

\item{choices}{Character vector of choices for the participant.
If unnamed, then these values will be used both for button IDs
and for button labels.
If named, then values will be used for button IDs and names
will be used for button labels.}

\item{arrange_vertically}{Whether to arrange the response buttons vertically
(the default) as opposed to horizontally.}
}
\description{
Creates an n-alternative forced-foced choice page.
}
