% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{new_results_section}
\alias{new_results_section}
\title{New results section}
\usage{
new_results_section(label)
}
\arguments{
\item{label}{Label to give the new results section.}
}
\description{
Returns a test element that initialises a new results section.
}
