% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{new_timeline}
\alias{new_timeline}
\title{New timeline}
\usage{
new_timeline(x = stop("x not supplied"), dict = NULL,
  default_lang = "EN")
}
\arguments{
\item{x}{Expression defining a series of test elements.
This expression will be evaluated once for every language
provided in the dictionary argument (\code{dict}).
A call to \code{\link{i18n}(key)}, where \code{key}
is a term defined in \code{dict},
will be translated to its definitions in each respective language.}

\item{dict}{Dictionary object as created by \code{\link{i18n_dict}$new}.}

\item{default_lang}{If no dictionary is supplied, then
\code{new_timeline()} assumes that the current language is
\code{default_lang}.}
}
\description{
Creates a new timeline.
Timelines allow tests to support multiple languages.
}
\note{
Debugging is difficult within \code{new_timeline()}
because of its underlying macro definition.
When building a test, we recommend defining small timelines
first and subsequently combining them.
This helps to narrow down the source of any errors.
}
