% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{p_id<-}
\alias{p_id<-}
\title{Sets participant ID}
\usage{
p_id(state) <- value
}
\arguments{
\item{state}{The participant's \code{state} object.}

\item{value}{The value to set the ID to.}
}
\description{
Sets the participant's ID.
}
