% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{page}
\alias{page}
\title{New page}
\usage{
page(ui, admin_ui = NULL, label = NULL, final = FALSE,
  get_answer = NULL, save_answer = FALSE, validate = NULL,
  on_complete = NULL, next_elt = TRUE)
}
\arguments{
\item{ui}{Page UI. Can be either a character scalar (e.g.
"Welcome to the test!") or an object of class "shiny.tag",
e.g. \code{shiny::tags$p("Welcome to the test!")}.}

\item{admin_ui}{Optional UI component for the admin panel.}

\item{label}{Page label (character scalar).}

\item{final}{Whether or not the page is the final page in the test.}

\item{get_answer}{Optional function for extracting the participant's
answer from the current page.
The argument list should include \code{...}, \code{input},
and optionally \code{state}, the participant's \code{state} object.
The function should extract the answer from the Shiny
\code{input} object and return it as its output.}

\item{save_answer}{Whether or not to save the answer.}

\item{validate}{Optional validation function.
The argument list should include \code{...},
and any of:
\code{answer}, the participant's most recent answer;
\code{state}, the participant's state object;
\code{input}, the current page's Shiny input object;
\code{opt}, the test's option list as created by \code{test_options()};
\code{session}, the current Shiny session object.
It should return \code{TRUE} for a successful validation;
for an unsuccessful validation, it should return either \code{FALSE}
or a character scalar error message.
If validation fails then the page will be refreshed, usually
to give the user a chance to revise their input.}

\item{on_complete}{Optional function to execute on leaving the page
(after successful validation).
The argument list should include \code{...},
and any of:
\code{state}, the participant's state object;
\code{answer}, the participant's most recent answer;
\code{input}, the current page's Shiny input object;
\code{session}, the current Shiny session object;
\code{opt}, the test's option list as created by \code{test_options()}.}

\item{next_elt}{(Logical scalar) Whether to go to the next element
in the timeline once this page is completed.
This will typically be \code{TRUE},
except for special cases such as pages that load
testing sessions from file (\code{\link{get_p_id}}).}
}
\description{
This is the most general way to create a psychTestR page.
}
