% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{pt_options}
\alias{pt_options}
\title{Alias for test_options()}
\usage{
pt_options(...)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{test_options}()}.}
}
\description{
Alias for \code{test_options()} (deprecated).
}
