% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{reactive_page}
\alias{reactive_page}
\title{Reactive page}
\usage{
reactive_page(fun, next_elt = TRUE)
}
\arguments{
\item{fun}{Function that returns a \code{page} argument.
This function's argument list must include \code{...}.
It should also include one or more named arguments that provide
information for generating the new page: potential arguments
include \code{state}, corresponding to the participant's
\code{state} object,
\code{answer}, which is the most recent participant response,
and \code{opt}, which is the test's option list as created by
\code{test_options}.
The function should always return an object of class \code{page}.}

\item{next_elt}{(Logical scalar) Whether to go to the next element
in the timeline once this page is completed.
This will typically be \code{TRUE},
except for special cases such as pages that load
testing sessions from file (\code{\link{get_p_id}}).}
}
\description{
Creates a reactive page.
A reactive page is defined by a function that is called once the participant
reaches a given location in the timeline.
This function must be idempotent, i.e. calling it multiple times should
have the same effect as calling it once).
}
