% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{register_next_results_section}
\alias{register_next_results_section}
\title{Register next results section}
\usage{
register_next_results_section(x, label)
}
\arguments{
\item{x}{Where should the new section be registered?
Methods exist for \code{results} objects and for \code{state} objects.}

\item{label}{Label for the new section.}
}
\description{
Register next results section
}
