% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results<-}
\alias{results<-}
\title{Set results}
\usage{
results(state) <- value
}
\arguments{
\item{state}{The participant's \code{state} object.}

\item{value}{Value to set the \code{results} slot to.}
}
\description{
Sets the \code{results} slot in \code{state}.
}
