% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results}
\alias{results}
\title{Access results}
\usage{
results(state)
}
\arguments{
\item{state}{The participant's \code{state} object.}
}
\description{
Accesses the \code{results} slot in \code{state}.
}
