% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{save_result}
\alias{save_result}
\title{Save result}
\usage{
save_result(place, label, value)
}
\arguments{
\item{place}{Place to save the results.
Methods exist for \code{results} objects and for \code{state} objects.}

\item{label}{Label for the new datum.}

\item{value}{Value of the new datum.}
}
\description{
Saves a new datum into the participant's results store.
}
