% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{save_results_to_disk}
\alias{save_results_to_disk}
\title{Save results to disk}
\usage{
save_results_to_disk(complete, state, opt, ...)
}
\arguments{
\item{complete}{Whether or not the participant
has now completed the test (scalar Boolean).}

\item{state}{The participant's \code{state} object.}

\item{opt}{Test options as created by \code{test_options()}.}

\item{...}{Further arguments are allowed but ignored (for back-compatibility).}

\item{complete}{Whether the participant completed the test.}
}
\description{
Saves the current participant's results to disk.
This function can be called e.g. within a code block.
}
