% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{set_global}
\alias{set_global}
\title{Set global variable}
\usage{
set_global(key, value, state)
}
\arguments{
\item{key}{The variable's key (character scalar).}

\item{value}{The value to set the variable to.}

\item{state}{The participant's \code{state} object.}
}
\description{
Sets a global variable for the participant's testing session.
}
