% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{set_local}
\alias{set_local}
\title{Set local variable}
\usage{
set_local(key, value, state)
}
\arguments{
\item{key}{The variable's key (character scalar).}

\item{value}{The value to set the variable to.}

\item{state}{The participant's \code{state} object.}
}
\description{
Sets a local variable for the participant's testing session.
}
