% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{set_url_param}
\alias{set_url_param}
\title{Set a URL parameter}
\usage{
set_url_param(key, value, session, state)
}
\arguments{
\item{key}{URL parameter key.}

\item{value}{URL parameter value.}

\item{session}{The participant's \code{session} object.}

\item{state}{The participant's \code{state} object.}
}
\description{
Sets a URL parameter.
}
