% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{skip_n_pages}
\alias{skip_n_pages}
\title{Intended to be called within e.g. code blocks to skip pages}
\usage{
skip_n_pages(state, n)
}
\description{
Intended to be called within e.g. code blocks to skip pages
}
