% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{test_options}
\alias{test_options}
\title{Test options}
\usage{
test_options(title, admin_password, researcher_email = NULL,
  max_num_participants = NULL, demo = FALSE, debug_locally = FALSE,
  log_error = TRUE, show_full_error_msg = TRUE, notify_error = FALSE,
  notify_new_participant = FALSE, pushbullet_email = NULL,
  pushbullet_apikey = NULL, max_participants_msg = NULL,
  server_closed_msg = NULL, problems_info = "default",
  theme = shinythemes::shinytheme("readable"), auto_p_id = TRUE,
  enable_resume_session = TRUE, allow_any_p_id_url = FALSE,
  force_p_id_from_url = FALSE, enable_admin_panel = TRUE,
  output_dir = "output", session_timeout_min = 7 * 24 * 60,
  clean_sessions_interval_min = 15, logo = NULL, logo_width = NULL,
  logo_height = NULL)
}
\arguments{
\item{title}{The test's title.}

\item{admin_password}{Password to access the admin panel.}

\item{researcher_email}{Researcher's email; used in participant help message.}

\item{max_num_participants}{Maximum number of participant completes
before the test automatically closes (excludes pilot participants
as indicated through the admin panel).}

\item{demo}{Whether the test is to be run in demo mode.}

\item{log_error}{Whether to log errors
(this feature is under development).}

\item{show_full_error_msg}{Whether to show full error messages.}

\item{notify_error}{Whether to send a Pushbullet message when
the test experiences an error
(this feature is under development).}

\item{notify_new_participant}{Whether to send a Pushbullet message when
a new participant completes the test.}

\item{pushbullet_email}{Email to send Pushbullet notifications to.}

\item{pushbullet_apikey}{Pushbullet API key (see https://www.pushbullet.com/).}

\item{max_participants_msg}{Message to display when the participant
quota is reached
(\code{NULL} gives default).}

\item{server_closed_msg}{Message to display when the server is closed
via the admin panel
(\code{NULL} gives default).}

\item{problems_info}{Message to display at the bottom of the screen
with advice about what to do if a problem occurs.
Defaults to a standard message including the researcher's email
(if provided).}

\item{theme}{Shiny theme: see e.g. the \code{shinythemes} package.}

\item{auto_p_id}{Whether or not to automatically generate an
ID for each participant.}

\item{enable_resume_session}{Whether to allow participants to resume
sessions by refreshing the page.}

\item{allow_any_p_id_url}{Whether to allow new participants
to give themselves an arbitrary participant ID by passing it
as a URL parameter.}

\item{force_p_id_from_url}{Whether to force new participants
to provide their participant ID as a URL parameter.}

\item{enable_admin_panel}{Wheter to enable the admin panel.}

\item{output_dir}{String identifying psychTestR's output directory.}

\item{session_timeout_min}{Minimum time until a participant's
session times out (minutes).}

\item{clean_sessions_interval_min}{How often should psychTestR
check for expired participant sessions (minutes).}

\item{logo}{Path to a logo to display in the header (optional).}

\item{logo_width}{Logo width, e.g. \code{"100px"}.}

\item{logo_height}{Logo height, e.g. \code{"50px"}.}

\item{debug}{Whether the test is to be debugged locally
(set to \code{TRUE} when developing locally).}
}
\description{
Defines the options for running a given test.
}
