% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{test_permissions}
\alias{test_permissions}
\title{Test write permissions}
\usage{
test_permissions(dir)
}
\arguments{
\item{dir}{Directory to test.}
}
\description{
Test the ability to write to a given directory.
}
