% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{text_input_page}
\alias{text_input_page}
\title{Text input page}
\usage{
text_input_page(label, prompt, one_line = TRUE, save_answer = TRUE,
  placeholder = NULL, button_text = "Next", width = "300px",
  height = "100px", validate = NULL, on_complete = NULL,
  admin_ui = NULL)
}
\arguments{
\item{label}{Label for the current page (character scalar).}

\item{prompt}{Prompt to display (character scalar or Shiny tag object)}

\item{save_answer}{See \code{\link{page}}.}

\item{placeholder}{Placeholder text for the text box (character scalar).}

\item{button_text}{Text for the submit button (character scalar).}

\item{width}{Width of the text box (character scalar, should be valid HTML).}

\item{height}{Height of the text box (character scalar, should be valid HTML).}

\item{validate}{See \code{\link{page}}.}

\item{on_complete}{See \code{\link{page}}.}

\item{admin_ui}{See \code{\link{page}}.}
}
\description{
Creates a page where the participant puts their
answer in a text box.
}
