% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\docType{data}
\name{timeline}
\alias{timeline}
\title{Timeline}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
timeline
}
\description{
Timelines are series of psychTestR test elements that chain
together to form a test.
They support internationalisation,
defining parallel series of test elements for the available languages.
}
\section{Creation}{

Timelines are created using \code{\link{new_timeline}}.
}

\section{Manipulation}{

Timelines can be combined with other timelines and with test elements
using \code{\link{c}}.
}

\section{Usage}{

Timelines are ultimately passed to \code{\link{make_test}}.
}

\section{Other methods}{

\code{x$get(language, i)} returns a list of test elements corresponding
to \code{language} as extracted from the timeline \code{x},
or, if \code{i} is not \code{NULL}, the ith such test element.

\code{x$drop_languages(drop)} removes support for a set of languages
from timeline \code{x}, where \code{drop} is the character vector
of languages to remove.
}

\keyword{datasets}
