% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{trigger_button}
\alias{trigger_button}
\title{Trigger button}
\usage{
trigger_button(inputId, label, icon = NULL, width = NULL,
  enable_after = 0, ...)
}
\arguments{
\item{inputId}{See \link[shiny]{actionButton}}

\item{label}{See \link[shiny]{actionButton}}

\item{icon}{See \link[shiny]{actionButton}}

\item{width}{See \link[shiny]{actionButton}}

\item{enable_after}{Number of seconds after which responses should be permitted.}

\item{...}{Additional arguments to pass to \link[shiny]{actionButton}.}
}
\description{
A version of \code{shiny::actionButton} that triggers
the next psychTestR page.
}
