% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{video_NAFC_page}
\alias{video_NAFC_page}
\title{Make NAFC video page}
\usage{
video_NAFC_page(prompt, choices, url, set_global = NULL,
  type = tools::file_ext(url), video_width = "100\%",
  arrange_choices_vertically = TRUE, wait = TRUE, loop = FALSE, ...)
}
\arguments{
\item{prompt}{Prompt to be displayed above the response choices.
Can be either a scalar character (e.g. "What is 2 + 2?")
or an object of class "shiny.tag", e.g. \code{shiny::tags$p("What is 2 + 2?")}.}

\item{choices}{Character vector of choices for the participant.
If unnamed, then these values will be used both for button IDs
and for button labels.
If named, then values will be used for button IDs and names
will be used for button labels.}

\item{url}{URL to the video.
Can be an absolute URL (e.g. "http://mysite.com/video.mp4")
or a URL relative to the /www directory (e.g. "video.mp4").}

\item{type}{Video type (e.g. 'mp4'). Defaults to the provided file extension.}

\item{video_width}{Video width, as passed to HTML (e.g. '50px').}

\item{arrange_choices_vertically}{Whether to arrange the response buttons vertically
(the default) as opposed to horizontally.}

\item{wait}{Whether to wait for the video to finish before displaying
the response buttons.}

\item{loop}{Whether the video should loop.}

\item{...}{Further parameters to be passed to \code{\link{page}}.}
}
\description{
Creates an n-alternative forced-foced choice page with a video prompt.
}
