% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{volume_calibration_page}
\alias{volume_calibration_page}
\title{Make volume calibration page}
\usage{
volume_calibration_page(url, type = tools::file_ext(url),
  prompt = NULL, button_text = "Next", on_complete = NULL,
  admin_ui = NULL, btn_play_prompt = "Click here to play")
}
\arguments{
\item{url}{URL to the audio.
Can be an absolute URL (e.g. "http://mysite.com/audio.mp3")
or a URL relative to the /www directory (e.g. "audio.mp3").}

\item{type}{Audio type (e.g. 'mp3'). Defaults to the provided file extension.}

\item{prompt}{Prompt to be displayed. If left \code{NULL},
a sensible English prompt is provided.}

\item{button_text}{Button text (character scalar).}

\item{on_complete}{Optional function to execute on leaving the page
(after successful validation).
The argument list should include \code{...},
and any of:
\code{state}, the participant's state object;
\code{answer}, the participant's most recent answer;
\code{input}, the current page's Shiny input object;
\code{session}, the current Shiny session object;
\code{opt}, the test's option list as created by \code{test_options()}.}

\item{admin_ui}{Optional UI component for the admin panel.}

\item{btn_play_prompt}{Text to display as a prompt for starting the audio.
Ordinarily the participant will not see this,
but it might appear if the internet connection is poor,
or if the participant refreshes the page.}
}
\description{
Creates a page for the participant to calibrate their volume,
using example audio, and the volume controls on their computer.
}
