% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{while_loop}
\alias{while_loop}
\title{While loop}
\usage{
while_loop(test, logic)
}
\arguments{
\item{test}{Test function to execute.
Argument list should include \code{...};
further permitted arguments are:
\code{state}, the participant's \code{state} object;
\code{input}, the Shiny \code{input} object for the current page;
\code{output}, the Shiny \code{output} object for the current page;
\code{session}, the Shiny \code{session} object;
\code{opt}, test options as created by \code{test_options()}.}

\item{logic}{List of psychTestR test elements.}
}
\description{
This function creates a "while loop" in the participant's testing
session. This "while loop" corresponds to the following procedure:
\enumerate{
\item Check whether a condition (termed \code{test}) is satisfied.
\item If the condition is satisfied,
execute a series of test elements (termed \code{logic}),
otherwise exit the while loop.
\item Loop back to step 1.
}
}
\note{
The previous version of this function, \code{loop_while},
always executed the series of test elements at least once,
even if the condition was never satisfied.
}
