% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{with_i18n_state}
\alias{with_i18n_state}
\title{With i18n state}
\usage{
with_i18n_state(dictionary = stop("dictionary not supplied"),
  language = stop("language not supplied"), x = stop("x not supplied"))
}
\arguments{
\item{dictionary}{An internationalisation dictionary as created by \code{\link{i18n_dict}}.}

\item{language}{A language code corresponding to a language supported in \code{dictionary}.}

\item{x}{Expression to evaluate.}
}
\value{
The result of evaluating \code{x} using the provided dictionary and language.
}
\description{
Sets the current internalisation state (\code{\link{I18N_STATE}})
to a provided dictionary and language,
and evaluates a provided expression.
This macro is called automatically within \code{\link{new_timeline}}
and most users will never need to call it directly.
}
