% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_tester.R
\name{AppTester}
\alias{AppTester}
\title{App tester}
\description{
This is an R6 class for running automated evaluations of psychTestR tests
using the \code{shinytest} framework.
When you create an \code{AppTester} object,
it spins up a Shiny app and a headless web browser,
which you can then interact with by calling the methods of the
\code{AppTester} object.
}
\details{
\itemize{
\item \code{app <- AppTester$new(dir)}:
creates a new \code{AppTester} object. The argument
\code{dir} should be a directory containing a file \code{app.R},
with \code{app.R} containing your psychTestR test.
\item \code{app$stop()} - stops the daemon and deletes the output directory.
Call this once you're finished with the \code{AppTester} object.
\item \code{app$click(buttons)} - clicks all the buttons with IDs
listed in the character vector \code{buttons}.
\item \code{app$click_next()} - clicks the "Next" button.
\item \code{app$get_ui()} - returns the HTML of the current UI of the psychTestR test.
\item \code{app$get_ui_text()} - extracts displayed text from \code{app$get_ui()}
using some heuristics. If \code{squish = TRUE}, extra whitespace
is removed.
\item \code{app$expect_ui_text(text, squish = TRUE)} - throws an error
if the output of \code{app$get_ui_text(squish = squish)}
is not equal to the argument \code{text}.
\item \code{app$get_locals()} - returns the current local variables.
\item \code{app$get_globals()} - returns the current global variables.
\item \code{app$get_title()} - returns the current title.
\item \code{app$expect_title(text)} - throws an error
if the output of \code{app$get_title()}
is not equal to the argument \code{text}.
\item \code{app$get_problems_info()} - returns the "problems info"
text displayed at the bottom of the page.
\item \code{app$expect_problems_info(text)} - throws an error
if the output of \code{app$get_problems_info()}
is not equal to the argument \code{text}.
\item \code{app$get_results()} - returns the current results for the current participant.
\item \code{app$expect_results(val)} - throws an error
if the output of \code{as.list(app$get_results())}
is not equal to \code{val}.
\item \code{app$clean_output()} - deletes the output directory.
\item \code{app$set_inputs(formA = "Text A", formB = "Text B")} -
passes (for example) the string \code{"Text A"} to the UI
input with ID \code{"formA"} and the string \code{"Text B"}
to the UI input with ID \code{"formB"}.
\item \code{...} - \code{AppTester} objects inherit from \code{\link[shinytest]{ShinyDriver}}
objects, and have access to all of their methods.
}
}
\note{
\code{AppTester} depends on the prior installation of a headless web browser, PhantomJS,
which you can install with \code{\link[shinytest]{installDependencies}}.
}
\section{Super class}{
\code{\link[shinytest:ShinyDriver]{shinytest::ShinyDriver}} -> \code{AppTester}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AppTester-get_ui}{\code{AppTester$get_ui()}}
\item \href{#method-AppTester-get_ui_text}{\code{AppTester$get_ui_text()}}
\item \href{#method-AppTester-str_squish}{\code{AppTester$str_squish()}}
\item \href{#method-AppTester-expect_ui_text}{\code{AppTester$expect_ui_text()}}
\item \href{#method-AppTester-get_locals}{\code{AppTester$get_locals()}}
\item \href{#method-AppTester-get_globals}{\code{AppTester$get_globals()}}
\item \href{#method-AppTester-get_title}{\code{AppTester$get_title()}}
\item \href{#method-AppTester-expect_title}{\code{AppTester$expect_title()}}
\item \href{#method-AppTester-get_problems_info}{\code{AppTester$get_problems_info()}}
\item \href{#method-AppTester-expect_problems_info}{\code{AppTester$expect_problems_info()}}
\item \href{#method-AppTester-get_results}{\code{AppTester$get_results()}}
\item \href{#method-AppTester-expect_results}{\code{AppTester$expect_results()}}
\item \href{#method-AppTester-stop}{\code{AppTester$stop()}}
\item \href{#method-AppTester-clean_output}{\code{AppTester$clean_output()}}
\item \href{#method-AppTester-set_inputs}{\code{AppTester$set_inputs()}}
\item \href{#method-AppTester-click}{\code{AppTester$click()}}
\item \href{#method-AppTester-click_next}{\code{AppTester$click_next()}}
\item \href{#method-AppTester-clone}{\code{AppTester$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="checkUniqueWidgetNames"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-checkUniqueWidgetNames'><code>shinytest::ShinyDriver$checkUniqueWidgetNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="enableDebugLogMessages"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-enableDebugLogMessages'><code>shinytest::ShinyDriver$enableDebugLogMessages()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="executeScript"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-executeScript'><code>shinytest::ShinyDriver$executeScript()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="executeScriptAsync"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-executeScriptAsync'><code>shinytest::ShinyDriver$executeScriptAsync()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="expectUpdate"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-expectUpdate'><code>shinytest::ShinyDriver$expectUpdate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="finalize"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-finalize'><code>shinytest::ShinyDriver$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="findElement"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-findElement'><code>shinytest::ShinyDriver$findElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="findElements"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-findElements'><code>shinytest::ShinyDriver$findElements()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="findWidget"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-findWidget'><code>shinytest::ShinyDriver$findWidget()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getAllValues"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-getAllValues'><code>shinytest::ShinyDriver$getAllValues()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getAppDir"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-getAppDir'><code>shinytest::ShinyDriver$getAppDir()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getAppFilename"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-getAppFilename'><code>shinytest::ShinyDriver$getAppFilename()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getDebugLog"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-getDebugLog'><code>shinytest::ShinyDriver$getDebugLog()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getEventLog"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-getEventLog'><code>shinytest::ShinyDriver$getEventLog()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getRelativePathToApp"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-getRelativePathToApp'><code>shinytest::ShinyDriver$getRelativePathToApp()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getSnapshotDir"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-getSnapshotDir'><code>shinytest::ShinyDriver$getSnapshotDir()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getSource"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-getSource'><code>shinytest::ShinyDriver$getSource()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getTestsDir"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-getTestsDir'><code>shinytest::ShinyDriver$getTestsDir()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getTitle"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-getTitle'><code>shinytest::ShinyDriver$getTitle()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getUrl"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-getUrl'><code>shinytest::ShinyDriver$getUrl()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getValue"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-getValue'><code>shinytest::ShinyDriver$getValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getWindowSize"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-getWindowSize'><code>shinytest::ShinyDriver$getWindowSize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="goBack"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-goBack'><code>shinytest::ShinyDriver$goBack()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="initialize"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-initialize'><code>shinytest::ShinyDriver$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="isRmd"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-isRmd'><code>shinytest::ShinyDriver$isRmd()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="listWidgets"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-listWidgets'><code>shinytest::ShinyDriver$listWidgets()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="logEvent"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-logEvent'><code>shinytest::ShinyDriver$logEvent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="refresh"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-refresh'><code>shinytest::ShinyDriver$refresh()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="sendKeys"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-sendKeys'><code>shinytest::ShinyDriver$sendKeys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="setInputs"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-setInputs'><code>shinytest::ShinyDriver$setInputs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="setValue"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-setValue'><code>shinytest::ShinyDriver$setValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="setWindowSize"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-setWindowSize'><code>shinytest::ShinyDriver$setWindowSize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="snapshot"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-snapshot'><code>shinytest::ShinyDriver$snapshot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="snapshotCompare"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-snapshotCompare'><code>shinytest::ShinyDriver$snapshotCompare()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="snapshotDownload"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-snapshotDownload'><code>shinytest::ShinyDriver$snapshotDownload()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="snapshotInit"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-snapshotInit'><code>shinytest::ShinyDriver$snapshotInit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="takeScreenshot"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-takeScreenshot'><code>shinytest::ShinyDriver$takeScreenshot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="uploadFile"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-uploadFile'><code>shinytest::ShinyDriver$uploadFile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="waitFor"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-waitFor'><code>shinytest::ShinyDriver$waitFor()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="waitForShiny"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-waitForShiny'><code>shinytest::ShinyDriver$waitForShiny()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="waitForValue"><a href='../../shinytest/html/ShinyDriver.html#method-ShinyDriver-waitForValue'><code>shinytest::ShinyDriver$waitForValue()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppTester-get_ui"></a>}}
\if{latex}{\out{\hypertarget{method-AppTester-get_ui}{}}}
\subsection{Method \code{get_ui()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$get_ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppTester-get_ui_text"></a>}}
\if{latex}{\out{\hypertarget{method-AppTester-get_ui_text}{}}}
\subsection{Method \code{get_ui_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$get_ui_text(squish = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppTester-str_squish"></a>}}
\if{latex}{\out{\hypertarget{method-AppTester-str_squish}{}}}
\subsection{Method \code{str_squish()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$str_squish(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppTester-expect_ui_text"></a>}}
\if{latex}{\out{\hypertarget{method-AppTester-expect_ui_text}{}}}
\subsection{Method \code{expect_ui_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$expect_ui_text(text, squish = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppTester-get_locals"></a>}}
\if{latex}{\out{\hypertarget{method-AppTester-get_locals}{}}}
\subsection{Method \code{get_locals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$get_locals()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppTester-get_globals"></a>}}
\if{latex}{\out{\hypertarget{method-AppTester-get_globals}{}}}
\subsection{Method \code{get_globals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$get_globals()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppTester-get_title"></a>}}
\if{latex}{\out{\hypertarget{method-AppTester-get_title}{}}}
\subsection{Method \code{get_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$get_title()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppTester-expect_title"></a>}}
\if{latex}{\out{\hypertarget{method-AppTester-expect_title}{}}}
\subsection{Method \code{expect_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$expect_title(text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppTester-get_problems_info"></a>}}
\if{latex}{\out{\hypertarget{method-AppTester-get_problems_info}{}}}
\subsection{Method \code{get_problems_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$get_problems_info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppTester-expect_problems_info"></a>}}
\if{latex}{\out{\hypertarget{method-AppTester-expect_problems_info}{}}}
\subsection{Method \code{expect_problems_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$expect_problems_info(text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppTester-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-AppTester-get_results}{}}}
\subsection{Method \code{get_results()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$get_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppTester-expect_results"></a>}}
\if{latex}{\out{\hypertarget{method-AppTester-expect_results}{}}}
\subsection{Method \code{expect_results()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$expect_results(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppTester-stop"></a>}}
\if{latex}{\out{\hypertarget{method-AppTester-stop}{}}}
\subsection{Method \code{stop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$stop(clean = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppTester-clean_output"></a>}}
\if{latex}{\out{\hypertarget{method-AppTester-clean_output}{}}}
\subsection{Method \code{clean_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$clean_output()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppTester-set_inputs"></a>}}
\if{latex}{\out{\hypertarget{method-AppTester-set_inputs}{}}}
\subsection{Method \code{set_inputs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$set_inputs(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppTester-click"></a>}}
\if{latex}{\out{\hypertarget{method-AppTester-click}{}}}
\subsection{Method \code{click()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$click(buttons)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppTester-click_next"></a>}}
\if{latex}{\out{\hypertarget{method-AppTester-click_next}{}}}
\subsection{Method \code{click_next()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$click_next()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppTester-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AppTester-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
