% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_tester.R
\name{AppTester}
\alias{AppTester}
\title{App tester}
\description{
This is an R6 class for running automated evaluations of psychTestR tests
using the \code{shinytest} framework.
When you create an \code{AppTester} object,
it spins up a Shiny app and a headless web browser,
which you can then interact with by calling the methods of the
\code{AppTester} object.
}
\details{
\itemize{
\item \code{app <- AppTester$new(dir)}:
creates a new \code{AppTester} object. The argument
\code{dir} should be a directory containing a file \code{app.R},
with \code{app.R} containing your psychTestR test.
\item \code{app$stop()} - stops the daemon and deletes the output directory.
Call this once you're finished with the \code{AppTester} object.
\item \code{app$click(buttons)} - clicks all the buttons with IDs
listed in the character vector \code{buttons}.
\item \code{app$click_next()} - clicks the "Next" button.
\item \code{app$get_ui()} - returns the HTML of the current UI of the psychTestR test.
\item \code{app$get_ui_text()} - extracts displayed text from \code{app$get_ui()}
using some heuristics. If \code{squish = TRUE}, extra whitespace
is removed.
\item \code{app$expect_ui_text(text, squish = TRUE)} - throws an error
if the output of \code{app$get_ui_text(squish = squish)}
is not equal to the argument \code{text}.
\item \code{app$get_locals()} - returns the current local variables.
\item \code{app$get_globals()} - returns the current global variables.
\item \code{app$get_title()} - returns the current title.
\item \code{app$expect_title(text)} - throws an error
if the output of \code{app$get_title()}
is not equal to the argument \code{text}.
\item \code{app$get_problems_info()} - returns the "problems info"
text displayed at the bottom of the page.
\item \code{app$expect_problems_info(text)} - throws an error
if the output of \code{app$get_problems_info()}
is not equal to the argument \code{text}.
\item \code{app$get_results()} - returns the current results for the current participant.
\item \code{app$expect_results(val)} - throws an error
if the output of \code{as.list(app$get_results())}
is not equal to \code{val}.
\item \code{app$clean_output()} - deletes the output directory.
\item \code{app$set_inputs(formA = "Text A", formB = "Text B")} -
passes (for example) the string \code{"Text A"} to the UI
input with ID \code{"formA"} and the string \code{"Text B"}
to the UI input with ID \code{"formB"}.
\item \code{...} - \code{AppTester} objects inherit from \code{\link[shinytest]{ShinyDriver}}
objects, and have access to all of their methods.
}
}
\note{
\code{AppTester} depends on the prior installation of a headless web browser, PhantomJS,
which you can install with \code{\link[shinytest]{installDependencies}}.
}
\section{Super class}{
\code{\link[shinytest:ShinyDriver]{shinytest::ShinyDriver}} -> \code{AppTester}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_ui}{\code{AppTester$get_ui()}}
\item \href{#method-get_ui_text}{\code{AppTester$get_ui_text()}}
\item \href{#method-str_squish}{\code{AppTester$str_squish()}}
\item \href{#method-expect_ui_text}{\code{AppTester$expect_ui_text()}}
\item \href{#method-get_locals}{\code{AppTester$get_locals()}}
\item \href{#method-get_globals}{\code{AppTester$get_globals()}}
\item \href{#method-get_title}{\code{AppTester$get_title()}}
\item \href{#method-expect_title}{\code{AppTester$expect_title()}}
\item \href{#method-get_problems_info}{\code{AppTester$get_problems_info()}}
\item \href{#method-expect_problems_info}{\code{AppTester$expect_problems_info()}}
\item \href{#method-get_results}{\code{AppTester$get_results()}}
\item \href{#method-expect_results}{\code{AppTester$expect_results()}}
\item \href{#method-stop}{\code{AppTester$stop()}}
\item \href{#method-clean_output}{\code{AppTester$clean_output()}}
\item \href{#method-set_inputs}{\code{AppTester$set_inputs()}}
\item \href{#method-click}{\code{AppTester$click()}}
\item \href{#method-click_next}{\code{AppTester$click_next()}}
\item \href{#method-clone}{\code{AppTester$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="checkUniqueWidgetNames">}\href{../../shinytest/html/ShinyDriver.html#method-checkUniqueWidgetNames}{\code{shinytest::ShinyDriver$checkUniqueWidgetNames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="enableDebugLogMessages">}\href{../../shinytest/html/ShinyDriver.html#method-enableDebugLogMessages}{\code{shinytest::ShinyDriver$enableDebugLogMessages()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="executeScript">}\href{../../shinytest/html/ShinyDriver.html#method-executeScript}{\code{shinytest::ShinyDriver$executeScript()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="executeScriptAsync">}\href{../../shinytest/html/ShinyDriver.html#method-executeScriptAsync}{\code{shinytest::ShinyDriver$executeScriptAsync()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="expectUpdate">}\href{../../shinytest/html/ShinyDriver.html#method-expectUpdate}{\code{shinytest::ShinyDriver$expectUpdate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="finalize">}\href{../../shinytest/html/ShinyDriver.html#method-finalize}{\code{shinytest::ShinyDriver$finalize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="findElement">}\href{../../shinytest/html/ShinyDriver.html#method-findElement}{\code{shinytest::ShinyDriver$findElement()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="findElements">}\href{../../shinytest/html/ShinyDriver.html#method-findElements}{\code{shinytest::ShinyDriver$findElements()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="findWidget">}\href{../../shinytest/html/ShinyDriver.html#method-findWidget}{\code{shinytest::ShinyDriver$findWidget()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getAllValues">}\href{../../shinytest/html/ShinyDriver.html#method-getAllValues}{\code{shinytest::ShinyDriver$getAllValues()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getAppDir">}\href{../../shinytest/html/ShinyDriver.html#method-getAppDir}{\code{shinytest::ShinyDriver$getAppDir()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getAppFilename">}\href{../../shinytest/html/ShinyDriver.html#method-getAppFilename}{\code{shinytest::ShinyDriver$getAppFilename()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getDebugLog">}\href{../../shinytest/html/ShinyDriver.html#method-getDebugLog}{\code{shinytest::ShinyDriver$getDebugLog()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getEventLog">}\href{../../shinytest/html/ShinyDriver.html#method-getEventLog}{\code{shinytest::ShinyDriver$getEventLog()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getRelativePathToApp">}\href{../../shinytest/html/ShinyDriver.html#method-getRelativePathToApp}{\code{shinytest::ShinyDriver$getRelativePathToApp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getSnapshotDir">}\href{../../shinytest/html/ShinyDriver.html#method-getSnapshotDir}{\code{shinytest::ShinyDriver$getSnapshotDir()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getSource">}\href{../../shinytest/html/ShinyDriver.html#method-getSource}{\code{shinytest::ShinyDriver$getSource()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getTestsDir">}\href{../../shinytest/html/ShinyDriver.html#method-getTestsDir}{\code{shinytest::ShinyDriver$getTestsDir()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getTitle">}\href{../../shinytest/html/ShinyDriver.html#method-getTitle}{\code{shinytest::ShinyDriver$getTitle()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getUrl">}\href{../../shinytest/html/ShinyDriver.html#method-getUrl}{\code{shinytest::ShinyDriver$getUrl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getValue">}\href{../../shinytest/html/ShinyDriver.html#method-getValue}{\code{shinytest::ShinyDriver$getValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="getWindowSize">}\href{../../shinytest/html/ShinyDriver.html#method-getWindowSize}{\code{shinytest::ShinyDriver$getWindowSize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="goBack">}\href{../../shinytest/html/ShinyDriver.html#method-goBack}{\code{shinytest::ShinyDriver$goBack()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="initialize">}\href{../../shinytest/html/ShinyDriver.html#method-initialize}{\code{shinytest::ShinyDriver$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="isRmd">}\href{../../shinytest/html/ShinyDriver.html#method-isRmd}{\code{shinytest::ShinyDriver$isRmd()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="listWidgets">}\href{../../shinytest/html/ShinyDriver.html#method-listWidgets}{\code{shinytest::ShinyDriver$listWidgets()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="logEvent">}\href{../../shinytest/html/ShinyDriver.html#method-logEvent}{\code{shinytest::ShinyDriver$logEvent()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="refresh">}\href{../../shinytest/html/ShinyDriver.html#method-refresh}{\code{shinytest::ShinyDriver$refresh()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="sendKeys">}\href{../../shinytest/html/ShinyDriver.html#method-sendKeys}{\code{shinytest::ShinyDriver$sendKeys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="setInputs">}\href{../../shinytest/html/ShinyDriver.html#method-setInputs}{\code{shinytest::ShinyDriver$setInputs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="setValue">}\href{../../shinytest/html/ShinyDriver.html#method-setValue}{\code{shinytest::ShinyDriver$setValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="setWindowSize">}\href{../../shinytest/html/ShinyDriver.html#method-setWindowSize}{\code{shinytest::ShinyDriver$setWindowSize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="snapshot">}\href{../../shinytest/html/ShinyDriver.html#method-snapshot}{\code{shinytest::ShinyDriver$snapshot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="snapshotCompare">}\href{../../shinytest/html/ShinyDriver.html#method-snapshotCompare}{\code{shinytest::ShinyDriver$snapshotCompare()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="snapshotDownload">}\href{../../shinytest/html/ShinyDriver.html#method-snapshotDownload}{\code{shinytest::ShinyDriver$snapshotDownload()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="snapshotInit">}\href{../../shinytest/html/ShinyDriver.html#method-snapshotInit}{\code{shinytest::ShinyDriver$snapshotInit()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="takeScreenshot">}\href{../../shinytest/html/ShinyDriver.html#method-takeScreenshot}{\code{shinytest::ShinyDriver$takeScreenshot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="uploadFile">}\href{../../shinytest/html/ShinyDriver.html#method-uploadFile}{\code{shinytest::ShinyDriver$uploadFile()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="waitFor">}\href{../../shinytest/html/ShinyDriver.html#method-waitFor}{\code{shinytest::ShinyDriver$waitFor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="waitForShiny">}\href{../../shinytest/html/ShinyDriver.html#method-waitForShiny}{\code{shinytest::ShinyDriver$waitForShiny()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytest" data-topic="ShinyDriver" data-id="waitForValue">}\href{../../shinytest/html/ShinyDriver.html#method-waitForValue}{\code{shinytest::ShinyDriver$waitForValue()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ui"></a>}}
\if{latex}{\out{\hypertarget{method-get_ui}{}}}
\subsection{Method \code{get_ui()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$get_ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ui_text"></a>}}
\if{latex}{\out{\hypertarget{method-get_ui_text}{}}}
\subsection{Method \code{get_ui_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$get_ui_text(squish = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-str_squish"></a>}}
\if{latex}{\out{\hypertarget{method-str_squish}{}}}
\subsection{Method \code{str_squish()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$str_squish(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-expect_ui_text"></a>}}
\if{latex}{\out{\hypertarget{method-expect_ui_text}{}}}
\subsection{Method \code{expect_ui_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$expect_ui_text(text, squish = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_locals"></a>}}
\if{latex}{\out{\hypertarget{method-get_locals}{}}}
\subsection{Method \code{get_locals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$get_locals()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_globals"></a>}}
\if{latex}{\out{\hypertarget{method-get_globals}{}}}
\subsection{Method \code{get_globals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$get_globals()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_title"></a>}}
\if{latex}{\out{\hypertarget{method-get_title}{}}}
\subsection{Method \code{get_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$get_title()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-expect_title"></a>}}
\if{latex}{\out{\hypertarget{method-expect_title}{}}}
\subsection{Method \code{expect_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$expect_title(text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_problems_info"></a>}}
\if{latex}{\out{\hypertarget{method-get_problems_info}{}}}
\subsection{Method \code{get_problems_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$get_problems_info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-expect_problems_info"></a>}}
\if{latex}{\out{\hypertarget{method-expect_problems_info}{}}}
\subsection{Method \code{expect_problems_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$expect_problems_info(text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-get_results}{}}}
\subsection{Method \code{get_results()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$get_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-expect_results"></a>}}
\if{latex}{\out{\hypertarget{method-expect_results}{}}}
\subsection{Method \code{expect_results()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$expect_results(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\if{latex}{\out{\hypertarget{method-stop}{}}}
\subsection{Method \code{stop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$stop(clean = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean_output"></a>}}
\if{latex}{\out{\hypertarget{method-clean_output}{}}}
\subsection{Method \code{clean_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$clean_output()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_inputs"></a>}}
\if{latex}{\out{\hypertarget{method-set_inputs}{}}}
\subsection{Method \code{set_inputs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$set_inputs(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-click"></a>}}
\if{latex}{\out{\hypertarget{method-click}{}}}
\subsection{Method \code{click()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$click(buttons)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-click_next"></a>}}
\if{latex}{\out{\hypertarget{method-click_next}{}}}
\subsection{Method \code{click_next()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$click_next()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppTester$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
