% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{audio_NAFC_page}
\alias{audio_NAFC_page}
\title{Make NAFC audio page}
\usage{
audio_NAFC_page(label, prompt, choices, url, type = tools::file_ext(url),
  on_complete = audio_NAFC_page.autosave(label, prompt, url),
  arrange_choices_vertically = TRUE, wait = TRUE, loop = FALSE, ...)
}
\arguments{
\item{prompt}{Prompt to be displayed above the response choices.
Can be either a scalar character (e.g. "What is 2 + 2?")
or an object of class "shiny.tag", e.g. \code{shiny::tags$p("What is 2 + 2?")}.}

\item{choices}{Character vector of choices for the participant.
If unnamed, then these values will be used both for button IDs
and for button labels.
If named, then values will be used for button IDs and names
will be used for button labels.}

\item{url}{URL to the audio
Can be an absolute URL (e.g. "http://mysite.com/audio.mp3")
or a URL relative to the /www directory (e.g. "audio.mp3").}

\item{type}{Audio type (e.g. 'mp3'). Defaults to the provided file extension.}

\item{arrange_choices_vertically}{Whether to arrange the response buttons vertically
(the default) as opposed to horizontally.}

\item{wait}{Whether to wait for the audio to finish before displaying
the response buttons.}

\item{loop}{Whether the audio should loop.}

\item{...}{Further parameters to be passed to \code{\link{page}}.}
}
\description{
Creates an n-alternative forced-foced choice page with an audio prompt.
}
