% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{audio_NAFC_page}
\alias{audio_NAFC_page}
\title{Make NAFC audio page}
\usage{
audio_NAFC_page(label, prompt, choices, url, labels = NULL,
  type = tools::file_ext(url), save_answer = TRUE,
  on_complete = NULL, arrange_choices_vertically = length(choices) >
  2L, wait = TRUE, loop = FALSE, admin_ui = NULL,
  btn_play_prompt = "Click here to play")
}
\arguments{
\item{label}{Label for the current page (character scalar).}

\item{prompt}{Prompt to be displayed above the response choices.
Can be either a character scalar (e.g. "What is 2 + 2?")
or an object of class "shiny.tag", e.g. \code{shiny::tags$p("What is 2 + 2?")}.}

\item{choices}{Character vector of choices for the participant.
If unnamed, then these values will be used both for button IDs
and for button labels.
If named, then values will be used for button IDs and names
will be used for button labels.}

\item{url}{URL to the audio
Can be an absolute URL (e.g. "http://mysite.com/audio.mp3")
or a URL relative to the /www directory (e.g. "audio.mp3").}

\item{labels}{Optional vector of labels for the NAFC choices.
If not \code{NULL}, will overwrite the names of \code{choices}.
This vector of labels can either be a character vector
or a list of Shiny tag objects, e.g. as created by \code{shiny::HTML()}.}

\item{type}{Audio type (e.g. 'mp3'). Defaults to the provided file extension.}

\item{save_answer}{Whether or not to save the answer.}

\item{on_complete}{Optional function to execute on leaving the page
(after successful validation).
The argument list should include \code{...},
and any of:
\code{state}, the participant's state object;
\code{answer}, the participant's most recent answer;
\code{input}, the current page's Shiny input object;
\code{session}, the current Shiny session object;
\code{opt}, the test's option list as created by \code{test_options()}.}

\item{arrange_choices_vertically}{Whether to arrange the response buttons vertically
(the default) as opposed to horizontally.}

\item{wait}{Whether to wait for the audio to finish before displaying
the response buttons.}

\item{loop}{Whether the audio should loop.}

\item{admin_ui}{Optional UI component for the admin panel.}

\item{btn_play_prompt}{Text to display as a prompt for starting the audio.
Ordinarily the participant will not see this,
but it might appear if the internet connection is poor,
or if the participant refreshes the page.}
}
\description{
Creates an n-alternative forced-foced choice page with an audio prompt.
}
