% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{code_block}
\alias{code_block}
\title{Code block}
\usage{
code_block(fun)
}
\arguments{
\item{fun}{Function to execute within code block.
This function's argument list must include \code{...}.
It may also include any of the following named arguments:
\code{state}, the participant's state object;
\code{opt}, which is the test's option list as created by \code{test_options()};
\code{input}, the current page's Shiny input object;
\code{output}, the current page's Shiny output object;
\code{session}, the current Shiny session object;
\code{elts}, the timeline (i.e. list of test elements).}
}
\description{
Creates a code block.
}
