% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{display_options}
\alias{display_options}
\title{Display options}
\usage{
display_options(full_screen = FALSE,
  content_background_colour = "white", content_border_width = "1px",
  show_header = TRUE, show_footer = TRUE, left_margin = 2L,
  right_margin = 2L, css = character())
}
\arguments{
\item{full_screen}{(Logical scalar)
Whether the display element for the primary test content
(corresponding to the \code{ui} slot of psychTestR pages)
should be expanded to fill the whole screen.
This is a convenience argument that works by overriding
the following arguments:
\itemize{
\item \code{content_border_width} (set to \code{"0px"})
\item \code{show_header} (set to \code{FALSE})
\item \code{show_footer} (set to \code{FALSE})
\item \code{left_margin} (set to \code{0L})
\item \code{right_margin} (set to \code{0L})
}}

\item{content_background_colour}{(Character scalar)
Background colour for the display element for the primary test content;
interpreted as a CSS expression.}

\item{content_border_width}{(Character scalar)
Border width for the display element for the primary test content;
interpreted as a CSS expression.}

\item{show_header}{(Logical scalar)
Whether the header should be shown (typically contains the title
and optionally a logo).}

\item{show_footer}{(Logical scalar)
Whether the footer should be shown (typically contains a link to the
admin panel).}

\item{left_margin}{(Integerish scalar)
Width of the left margin, with 0 corresponding to no margin
and 12 corresponding to the full screen width.}

\item{right_margin}{(Integerish scalar)
Width of the right margin, with 0 corresponding to no margin
and 12 corresponding to the full screen width.}

\item{css}{Character vector of file paths to CSS files for inclusion within the
psychTestR test. CSS files are used for styling HTML.
These file paths should be expressed relative to R's working directory}
}
\value{
A list for passing to the \code{display} argument of \code{test_options}.
}
\description{
Creates a list of display options that may be passed to
\code{\link{test_options}}.
}
