% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{dropdown_page}
\alias{dropdown_page}
\title{Make dropdown list page}
\usage{
dropdown_page(label, prompt, choices, alternative_choice = FALSE,
  alternative_text = "Other (please state)", save_answer = TRUE,
  validate = dropdown_page.validate(alternative_choice,
  alternative_text), on_complete = NULL, next_button_text = "Next",
  max_width_pixels = 200, admin_ui = NULL)
}
\arguments{
\item{label}{Page label (character scalar).}

\item{prompt}{Prompt to be displayed above the response choices.
Can be either a character scalar (e.g. "What is 2 + 2?")
or an object of class "shiny.tag", e.g. \code{shiny::tags$p("What is 2 + 2?")}.}

\item{choices}{Character vector of choices for the participant.
If names are provided, then these names will be used for display,
whereas the values will be stored in the results.}

\item{alternative_choice}{Whether or not to give the participant
the option of providing a free-text response instead of
selecting one of the dropdown options.}

\item{alternative_text}{Prompt for the free-text box (only relevant
if \code{alternative_choice} is set to \code{TRUE}).}

\item{save_answer}{Whether or not to save the answer.}

\item{validate}{Optional validation function.
The argument list should include \code{...},
and any of:
\code{answer}, the participant's most recent answer;
\code{state}, the participant's state object;
\code{input}, the current page's Shiny input object;
\code{opt}, the test's option list as created by \code{test_options()};
\code{session}, the current Shiny session object.
It should return \code{TRUE} for a successful validation;
for an unsuccessful validation, it should return either \code{FALSE}
or a character scalar error message.
If validation fails then the page will be refreshed, usually
to give the user a chance to revise their input.}

\item{on_complete}{Optional function to execute on leaving the page
(after successful validation).
The argument list should include \code{...},
and any of:
\code{state}, the participant's state object;
\code{answer}, the participant's most recent answer;
\code{input}, the current page's Shiny input object;
\code{session}, the current Shiny session object;
\code{opt}, the test's option list as created by \code{test_options()}.}

\item{next_button_text}{Text to display on the next-page button
(character scalar).}

\item{max_width_pixels}{Maximum width of the response UI, in pixels.}

\item{admin_ui}{Optional UI component for the admin panel.}
}
\description{
Creates a page where the response is to be selected from a dropdown list.
}
