% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{end_module}
\alias{end_module}
\title{End module}
\usage{
end_module()
}
\arguments{
\item{label}{Module label (scalar character).}
}
\description{
Returns a code block that ends a psychTestR module.
Modules have their own set of local variables.
They also have identifying labels that are stored alongside
result entries created during the module.
}
