% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{final_page}
\alias{final_page}
\title{New final page}
\usage{
final_page(body, admin_ui = NULL)
}
\arguments{
\item{body}{Page body. Can be either a scalar character (e.g.
"Welcome to the test!") or an object of class "shiny.tag",
e.g. \code{shiny::tags$p("Welcome to the test!")}.}

\item{admin_ui}{See \code{\link{page}}.}
}
\description{
Creates a page that concludes the test.
}
