% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{get_p_id}
\alias{get_p_id}
\title{Get participant ID}
\usage{
get_p_id(prompt = "Please enter your participant ID.",
  placeholder = "e.g. 10492817", button_text = "Next",
  width = "300px", validate = "auto", admin_ui = NULL)
}
\arguments{
\item{prompt}{Prompt to display (character scalar or Shiny tag object)}

\item{placeholder}{Placeholder text for the text box (character scalar).}

\item{button_text}{Text for the submit button (character scalar).}

\item{width}{Width of the text box (character scalar, should be valid HTML).}

\item{validate}{See \code{\link{page}}.}

\item{admin_ui}{See \code{\link{page}}.}
}
\description{
A psychTestR page that gets the participant to enter their ID.
}
\note{
Participant IDs must be between 1 and 100 characters long,
and solely comprise alphanumeric characters and underscores.
}
