% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{get_session_info}
\alias{get_session_info}
\title{Get session info}
\usage{
get_session_info(state, complete)
}
\value{
List containing session information.
}
\description{
Gets information about the current participant's session,
including things like participant ID, time started, and number of restarts.
}
