% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{i18n_dict}
\alias{i18n_dict}
\title{Dictionary}
\description{
Dictionaries allow psychTestR tests to
support multiple languages during test administration.
These dictionaries (class name: \code{i18n_dict})
define mappings between string keys (e.g. \code{prompt1})
and (potentially) multiple outputs,
each corresponding to different languages (e.g. English, German).
The dictionary is passed as an argument to \code{\link{new_timeline}},
and provides the context for the evaluation of calls to
\code{\link{i18n}}, responsible for translating individual terms.
}
\section{Methods}{

\code{i18n_dict$new(x, markdown)} defines and returns a new dictionary
which can be saved as an object.
Its first argument, \code{x}, should be a data frame defining the dictionary.
Each row of this data frame should correspond to a term to be translated.
It should have a column entitled 'key', which defines the identifying
keys for each term.
All other columns should provide translations into different languages,
with the language being identified by the column name,
according to ISO 639-2 conventions.
The second argument, \code{markdown}, is a scalar Boolean (default = \code{TRUE})
that determines whether the data frame is formatted in
Markdown or not (Markdown allows for efficient and readable
text markup, e.g. italic and bold).
Text is parsed according to standard Markdown conventions,
with one addition: two successive backslashes
(written "\\" in a text file, or "\\\\" in R)
are interpreted as a new paragraph.

\code{x$as.data.frame}, where \code{x} is a dictionary object,
converts the dictionary back to a data frame representation.

\code{x$edit(key, language, new)},
edits a given entry in the dictionary for a given \code{key} and \code{language},
replacing it with the value \code{new}.
See the vignette "Editing dictionaries" for details.

\code{x$translate(key, language, allow_missing)}
uses the dictionary \code{x}
to translate \code{key} into \code{language}.
If \code{allow_missing} is \code{TRUE}, then \code{NULL} is returned
when no translation is found, otherwise an error is thrown.
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{i18n_dict$new()}}
\item \href{#method-as.data.frame}{\code{i18n_dict$as.data.frame()}}
\item \href{#method-translate}{\code{i18n_dict$translate()}}
\item \href{#method-edit}{\code{i18n_dict$edit()}}
\item \href{#method-print}{\code{i18n_dict$print()}}
\item \href{#method-clone}{\code{i18n_dict$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{i18n_dict$new(x, markdown = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as.data.frame"></a>}}
\if{latex}{\out{\hypertarget{method-as.data.frame}{}}}
\subsection{Method \code{as.data.frame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{i18n_dict$as.data.frame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-translate"></a>}}
\if{latex}{\out{\hypertarget{method-translate}{}}}
\subsection{Method \code{translate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{i18n_dict$translate(key, language, allow_missing = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edit"></a>}}
\if{latex}{\out{\hypertarget{method-edit}{}}}
\subsection{Method \code{edit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{i18n_dict$edit(key, language, new)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{i18n_dict$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{i18n_dict$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
