% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{make_ui_NAFC}
\alias{make_ui_NAFC}
\title{Make NAFC buttons}
\usage{
make_ui_NAFC(choices, labels = NULL, hide = FALSE,
  arrange_vertically = length(choices) > 2L, style = "", id = "response_ui")
}
\arguments{
\item{choices}{Character vector of choices for the participant.
If unnamed, then these values will be used both for button IDs
and for button labels.
If named, then values will be used for button IDs and names
will be used for button labels.}

\item{labels}{Optional vector of labels for the NAFC choices.
If not \code{NULL}, will overwrite the names of \code{choices}.
This vector of labels can either be a character vector
or a list of Shiny tag objects, e.g. as created by \code{shiny::HTML()}.}

\item{hide}{Whether the response buttons should be hidden
(possibly to be shown later).}

\item{arrange_vertically}{Whether to arrange the response buttons vertically
(the default) as opposed to horizontally.}

\item{style}{CSS style information (character scalar).}

\item{id}{HTML ID for the div containing the response buttons.}
}
\description{
Creates HTML code for n-alternative forced-choice response options.
}
