% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{module}
\alias{module}
\title{Create module}
\usage{
module(label, ...)
}
\arguments{
\item{label}{Label for the module; should ideally be limited
to alphanumeric characters and underscores, but this is not essential.}

\item{...}{The psychTestR test elements that will constitute the module.}
}
\description{
Creates a psychTestR module.
A module encapsulates a portion of a psychTestR timeline,
providing it with its own environment of local variables.
Entering a module begins a new section in the psychTestR results object.
Modules are useful for wrapping particular tests or paradigms into
distributable units.
}
\seealso{
This function wraps the low-level functions
\code{\link{begin_module}} and \code{\link{end_module}}.
}
