% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{one_button_page}
\alias{one_button_page}
\title{New one-button page}
\usage{
one_button_page(body, admin_ui = NULL, button_text = "Next",
  on_complete = NULL)
}
\arguments{
\item{body}{Page body. Can be either a character scalar (e.g.
"Welcome to the test!") or an object of class "shiny.tag",
e.g. \code{shiny::tags$p("Welcome to the test!")}.}

\item{admin_ui}{See \code{\link{page}}.}

\item{button_text}{Text to display on the button.
Should be a character scalar.}

\item{on_complete}{See \code{\link{page}}.}
}
\description{
Creates a page with a prompt and one button which, when clicked,
advances to the next page.
This is typically used for giving the participant information
about the test.
}
