% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_elements.R
\name{one_button_page}
\alias{one_button_page}
\title{New one-button page}
\usage{
one_button_page(body, button_text = "Next", ...)
}
\arguments{
\item{body}{Page body. Can be either a scalar character (e.g.
"Welcome to the test!") or an object of class "shiny.tag",
e.g. \code{shiny::tags$p("Welcome to the test!")}.}

\item{button_text}{Text to display on the button.
Should be a scalar character vector.}

\item{...}{Further parameters to be passed to \code{\link{page}}.}
}
\description{
Creates a page with a prompt and one button which, when clicked,
advances to the next page.
This is typically used for giving the participant information
about the test.
}
