% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{page}
\alias{page}
\title{New page}
\usage{
page(ui, final = FALSE, on_complete = NULL, validate = NULL)
}
\arguments{
\item{ui}{Page UI. Can be either a scalar character (e.g.
"Welcome to the test!") or an object of class "shiny.tag",
e.g. \code{shiny::tags$p("Welcome to the test!")}.}

\item{final}{Whether or not the page is the final page in the test.}

\item{on_complete}{Optional function to run when page completes.
Should take the following form: function(state, input) ...
where \code{state} is the app's persistent state (a Shiny reactiveValues object)
and \code{input} is the input produced by the page UI.}

\item{validate}{Optional validation function.
Should take the same form as \code{on_complete}, and return
\code{TRUE} for a successful validation and \code{FALSE}
for an unsuccessful validation.
If validation fails then the page will be refreshed, usually
to give the user a chance to revise their input.}
}
\description{
This is the most general way to create a psychTest page.
}
