% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{reactive_page}
\alias{reactive_page}
\title{Careful - function must be idempotent (calling it multiple times should
have the same effect as calling it once).}
\usage{
reactive_page(fun)
}
\description{
Careful - function must be idempotent (calling it multiple times should
have the same effect as calling it once).
}
