% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_result}
\alias{save_result}
\title{#' @export
new_result <- function(key, data) {
  x <- list(key = key, data = data)
  class(x) <- "result"
  x
}}
\usage{
save_result(place, key, value)
}
\description{
#' @export
new_result <- function(key, data) {
  x <- list(key = key, data = data)
  class(x) <- "result"
  x
}
}
