% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{save_result}
\alias{save_result}
\title{#' @export
get_next_results_section <- function(x) {
  UseMethod("get_next_results_section")
}
#' @export
get_next_results_section.state <- function(x) {
  get_next_results_section(results(x))
}}
\usage{
save_result(place, label, value, metadata)
}
\description{
get_next_results_section.results <- function(x) {
  y <- attr(x, "new_section")
  z <- names(results)
  if (!is.null(y)) return(y)
  if (length(z) > 0) return(z[length(z)])
stop()
  else {
    names(results)
  }
}
}
