% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{test_permissions}
\alias{test_permissions}
\title{#' @export
delete_default_session_dirs <- function() {
  for (dir in default_session_dirs) {
    unlink(dir, recursive = TRUE)
  }
}}
\usage{
test_permissions(dir)
}
\description{
#' @export
delete_default_session_dirs <- function() {
  for (dir in default_session_dirs) {
    unlink(dir, recursive = TRUE)
  }
}
}
