% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{text_input_page}
\alias{text_input_page}
\title{Text input page}
\usage{
text_input_page(
  label,
  prompt,
  one_line = TRUE,
  save_answer = TRUE,
  placeholder = NULL,
  button_text = "Next",
  width = "300px",
  height = "100px",
  validate = NULL,
  on_complete = NULL,
  admin_ui = NULL
)
}
\arguments{
\item{label}{Label for the current page (character scalar).}

\item{prompt}{Prompt to display (character scalar or Shiny tag object).}

\item{one_line}{Whether the answer box only has one line of text.}

\item{save_answer}{Whether or not to save the answer.}

\item{placeholder}{Placeholder text for the text box (character scalar).}

\item{button_text}{Text for the submit button (character scalar).}

\item{width}{Width of the text box (character scalar, should be valid HTML).}

\item{height}{Height of the text box (character scalar, should be valid HTML).}

\item{validate}{Optional validation function.
The argument list should include \code{...},
and any of:
\code{answer}, the participant's most recent answer;
\code{state}, the participant's state object;
\code{input}, the current page's Shiny input object;
\code{opt}, the test's option list as created by \code{test_options()};
\code{session}, the current Shiny session object.
It should return \code{TRUE} for a successful validation;
for an unsuccessful validation, it should return either \code{FALSE}
or a character scalar error message.
If validation fails then the page will be refreshed, usually
to give the user a chance to revise their input.}

\item{on_complete}{Optional function to execute on leaving the page
(after successful validation).
The argument list should include \code{...},
and any of:
\code{state}, the participant's state object;
\code{answer}, the participant's most recent answer;
\code{input}, the current page's Shiny input object;
\code{session}, the current Shiny session object;
\code{opt}, the test's option list as created by \code{test_options()}.}

\item{admin_ui}{Optional UI component for the admin panel.}
}
\description{
Creates a page where the participant puts their
answer in a text box.
}
