% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{trigger_button}
\alias{trigger_button}
\title{Trigger button}
\usage{
trigger_button(inputId, label, icon = NULL, width = NULL,
  enable_after = 0, ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{icon}{An optional \code{\link[=icon]{icon()}} to appear on the button.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[=validateCssUnit]{validateCssUnit()}}.}

\item{enable_after}{Number of seconds after which responses should be permitted.}

\item{...}{Named attributes to be applied to the button or link.}
}
\description{
A version of \code{shiny::actionButton} that triggers
the next psychTestR page.
}
