% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{volume_calibration_page}
\alias{volume_calibration_page}
\title{Make volume calibration page}
\usage{
volume_calibration_page(url, type = tools::file_ext(url),
  prompt = NULL, button_text = "Next", on_complete = NULL,
  admin_ui = NULL)
}
\arguments{
\item{url}{URL to the audio
Can be an absolute URL (e.g. "http://mysite.com/audio.mp3")
or a URL relative to the /www directory (e.g. "audio.mp3").}

\item{type}{Audio type (e.g. 'mp3'). Defaults to the provided file extension.}

\item{prompt}{Prompt to be displayed. If left \code{NULL},
a sensible English prompt is provided.}
}
\description{
Creates a page for the participant to calibrate their volume,
using example audio, and the volume controls on their computer.
}
